jarvis/brain/requirements.txt
anthropic>=0.40.0
faster-whisper>=1.0.3
piper-tts>=1.2.0
websockets>=12.0
python-dotenv>=1.0.1
paramiko>=3.4.0
httpx>=0.27.0
pyyaml>=6.0.2
structlog>=24.4.0
tenacity>=9.0.0
